skills/transcribe-pdf/scripts/requirements.txt
PyMuPDF==1.26.7
Pillow==10.4.0

